// src/types/TransportAttributes.ts
export interface TransportAttributes {
    id: string;
    model?: string;
    year_release?: string;
    id_vehicle_type?: number;
    id_engine_type?: number;
    id_organization?: number;
    id_transport?: string; 
    block_id?: string;
    is_archived?: boolean;
    received_count?: number;
}